15/02/2021
Fixed the main and off hand slot so it will now dissappear when empty. Added to select wheel so if the slot is empty, the player cant access it. Added some simply text under the main and off hand slot to indicate what buttons will be used. Fixed the tooltip so it will go to the mouse's location. Adjusted the slots on the select wheel. Added code to disable the equip button on collectables.